+++ src/assets/Dashboard.js
@@ -1,24 +1,23 @@
 import React, { useEffect } from 'react';
+import { useRouteMatch, Switch, Route } from 'react-router-dom';
 import { connect } from 'react-redux';
 import {
   fetchRooms,
   fetchRecent,
   fetchUser,
   retrieveFullSearchResults,
-  setSearch
+  setSearch,
 } from '../store/actions';
-import {
-  Layout,
-  Breadcrumb,
-} from 'antd';
+import { Layout, Breadcrumb } from 'antd';
 import Navbar from './Navbar';
 import SiderMenu from './SiderMenu';
-import Feed from './Feed';
+import DashboardContent from './DashboardContent';
+import Room from './Room';
 
 const Dashboard = (props) => {
+  const { path } = useRouteMatch();
 
   const { Header, Content, Sider } = Layout;
-
   useEffect(() => {
     props.fetchRooms();
     // props.fetchRecent(); // This is throwing an internal server error
@@ -28,33 +27,41 @@ const Dashboard = (props) => {
   }, []);
 
   return (
-    <Layout>
-      <Header className="header" style={{ width: '100%' }}>
-        <Navbar/>
-      </Header>
+    <>
       <Layout>
-        <Sider width={200} className="site-layout-background">
-          <SiderMenu />
-        </Sider>
-        <Layout style={{ padding: '0 24px 24px' }}>
-          <Breadcrumb style={{ margin: '16px 0' }}>
-            <Breadcrumb.Item>Home</Breadcrumb.Item>
-            <Breadcrumb.Item>List</Breadcrumb.Item>
-            <Breadcrumb.Item>App</Breadcrumb.Item>
-          </Breadcrumb>
-          <Content
-            className="site-layout-background"
-            style={{
-              padding: 24,
-              margin: 0,
-              minHeight: 280,
-            }}
-          >
-            <Feed/>
-          </Content>
+        <Header className="header" style={{ width: '100%' }}>
+          <Navbar />
+        </Header>
+        <Layout>
+          <Sider width={200} className="site-layout-background">
+            <SiderMenu />
+          </Sider>
+          <Layout style={{ padding: '0 24px 24px' }}>
+            <Breadcrumb style={{ margin: '16px 0' }}>
+              <Breadcrumb.Item>Home</Breadcrumb.Item>
+              <Breadcrumb.Item>List</Breadcrumb.Item>
+              <Breadcrumb.Item>App</Breadcrumb.Item>
+            </Breadcrumb>
+            <Content
+              className="site-layout-background"
+              style={{
+                padding: 24,
+                margin: 0,
+                minHeight: 280,
+              }}
+            >
+              <DashboardContent />
+            </Content>
+          </Layout>
         </Layout>
       </Layout>
-    </Layout>
+
+      <Switch>
+        <Route path={`${path}/room/:id`}>
+          <Room />
+        </Route>
+      </Switch>
+    </>
   );
 };
 
@@ -72,5 +79,5 @@ export default connect(mapStateToProps, {
   fetchRecent,
   fetchUser,
   retrieveFullSearchResults,
-  setSearch
+  setSearch,
 })(Dashboard);
