+++ src/assets/Room.js
@@ -1,8 +1,7 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
-import { useParams, Link } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import {
-  fetchRecent,
   fetchPostByRoom,
   fetchRooms,
   setDrawerVisibility,
